requirements.txt
Flask==3.1.0 
pywebview==5.3.2
psutil==6.1.1


